add image functionality
This commit uses the unsplash-js module to finish the last two
requirements of this angular project. The images are loaded on component
init, and the user can select an image from the modal. The user can then
reopen the modal and select a new image or remove the selected image
entirely by clicking the "X" button.

This commit also updates the README.